Clean up Task test factories
