+++ src/test/kotlin/logic/usecases/task/DeleteTaskUseCaseTest.kt
@@ -1,7 +1,7 @@
 package logic.usecases.task
 
 import com.google.common.truth.Truth.assertThat
-import helpers.DeleteTaskTestFactory
+import helper.TaskFactory.TASK_1
 import io.mockk.coEvery
 import io.mockk.coVerify
 import io.mockk.mockk
@@ -28,7 +28,7 @@ class DeleteTaskUseCaseTest {
     @Test
     fun `should call deleteTask when parameters is correct`() =runTest {
         // Given
-        val task = DeleteTaskTestFactory.TASK_1
+        val task = TASK_1
         // When
         useCase.execute(task.id.toString())
 
@@ -38,7 +38,7 @@ class DeleteTaskUseCaseTest {
     @Test
     fun `should returns true when task is deleted successfully`() =runTest {
         // Given
-        val task = DeleteTaskTestFactory.TASK_1
+        val task = TASK_1
         coEvery { repository.deleteTask(task.id) } returns true
         // When
         val result  = useCase.execute(task.id.toString())
@@ -52,7 +52,7 @@ class DeleteTaskUseCaseTest {
     @Test
     fun `should returns false when task is not deleted`() =runTest {
         // Given
-        val task = DeleteTaskTestFactory.TASK_1
+        val task = TASK_1
         coEvery { repository.deleteTask(task.id) } returns false
         // When
         val result  = useCase.execute(task.id.toString())
